fix(entropy): validate two unchecked sub-streams; widen a u8 cluster count
Three findings from auditing the decoder alongside the SIMD-tier work. All
three were verified against source before changing anything, and one of the
reported claims did not survive that check — see below.

1. Two entropy sub-streams were never validated with `check_final_state`: the
   patches dictionary (features/patches.rs) and the permuted TOC
   (headers/encodings.rs). That check verifies the symbol reader's deferred
   errors, the bit reader's, and the ANS final-state checksum. Every other
   sub-stream in the decoder already ran it — splines, the modular tree,
   coefficient orders, ICC, the context map — as does libjxl.

   The reported claim was that a corrupt TOC permutation 'silently reorders
   sections'. That is CORRECTED, not confirmed. `Permutation::decode` does
   bounds-check every Lehmer code, so a corrupt stream yields a valid-but-wrong
   permutation as reported — but it is not silent: a randomized search of
   200,000 one-to-three-byte corruptions per fixture across has_permutation,
   has_permutation_with_container and both grayscale_patches fixtures, plus an
   exhaustive single-bit sweep of has_permutation.jxl, produced 160 inputs that
   the new checks reject and ZERO that were accepted without them. Downstream
   section-length and bit-padding validation caught every one. So this is
   hardening and parity, not a silent-corruption fix: corruption is now caught
   at the sub-stream carrying it, with AnsChecksumMismatch, rather than
   incidentally later.

   Mutation-verified: with the permutation check removed, the new test's four
   corrupted inputs fail with SectionTooShort / NonZeroPadding instead of
   AnsChecksumMismatch. No test is added for the patches check, because no
   input was found whose behaviour it changes; it is source-verified parity
   with libjxl and with every sibling site.

2. `Histograms::decode` computed `num_histograms` as
   `*context_map.iter().max().unwrap() + 1` in u8, while both sibling sites
   (`Histograms::num_histograms`, `verify_context_map`) widen first. Cluster
   index 255 is legal — `decode_context_map` rejects only values above
   u8::MAX, and `verify_context_map` requires every value below the maximum to
   be present, so a 256-cluster map is constructible — and 255u8 + 1 overflows:
   a panic under debug assertions, a wrap to zero in release, which left
   `uint_configs` empty and the codes decoded for no histograms at all.

3. The public `coalescing` option is inert: four occurrences, all in
   api/options.rs, zero reads. The decoder always coalesces, so `true` is
   honest and `false` is silently ignored. Removing a public field is a
   breaking change and this crate batches those, so it is now documented as
   unimplemented, queued in the changelog's QUEUED BREAKING CHANGES, and
   tracked in #57 — rather than left advertised.

Decoded output is unchanged: all 80 fixtures decode to byte-identical bytes
against base cd2d6f7 on this host.

+++ zenjxl-decoder/src/entropy_coding/decode.rs
@@ -635,18 +635,22 @@ impl Histograms {
         } else {
             br.read(2)? as usize + 5
         };
-        let num_histograms = *context_map.iter().max().unwrap() + 1;
+        // Widen before adding: `context_map` is `Vec<u8>`, and a cluster index
+        // of 255 is legal — `decode_context_map` only rejects values above
+        // `u8::MAX`, and `verify_context_map` requires every value below the
+        // maximum to be present, so a 256-cluster map is constructible. In u8
+        // that `+ 1` overflowed: a panic under debug assertions, and a wrap to
+        // zero in release, which left `uint_configs` empty and the codes
+        // decoded for no histograms at all. Both sibling sites already widen
+        // first (`Histograms::num_histograms`, `verify_context_map`).
+        let num_histograms = *context_map.iter().max().unwrap() as usize + 1;
         let uint_configs = ((0..num_histograms).map(|_| HybridUint::decode(log_alpha_size, br)))
             .collect::<Result<_, Error>>()?;
 
         let codes = if use_prefix_code {
-            Codes::Huffman(HuffmanCodes::decode(num_histograms as usize, br)?)
+            Codes::Huffman(HuffmanCodes::decode(num_histograms, br)?)
         } else {
-            Codes::Ans(AnsCodes::decode(
-                num_histograms as usize,
-                log_alpha_size,
-                br,
-            )?)
+            Codes::Ans(AnsCodes::decode(num_histograms, log_alpha_size, br)?)
         };
 
         Ok(Histograms {
